Class structure modification
